Menu refactoring
