+++ include/menus/cheats_menu.h
@@ -1,0 +1,7 @@
+#include "menu.h"
+
+class CheatsMenu : public Menu {
+public:
+    CheatsMenu() : Menu() {}
+    static void render(Font& font);
+};
