+++ include/menus/main_menu.h
@@ -1,0 +1,19 @@
+#include "menu.h"
+
+enum MainMenuIndex {
+    CHEAT_INDEX,
+    FLAGS_INDEX,
+    INVENTORY_INDEX,
+    MEMORY_INDEX,
+    PRACTICE_INDEX,
+    SCENE_INDEX,
+    SETTINGS_INDEX,
+    TOOLS_INDEX,
+    WARPING_INDEX,
+};
+
+class MainMenu : public Menu {
+public:
+    MainMenu() : Menu() {}
+    static void render(Font& font);
+};
