+++ include/menus/memory_editor_menu.h
@@ -1,0 +1,12 @@
+#include "menu.h"
+
+struct MemoryLine {
+    bool line_selected = false;
+    bool value_selected = false;
+};
+
+class MemoryEditorMenu : public Menu {
+public:
+    MemoryEditorMenu() : Menu() {}
+    static void render(Font& font);
+};
