+++ src/gz_flags.cpp
@@ -5,7 +5,9 @@
 #include "gz_flags.h"
 #include "gorge.h"
 #include "rollcheck.h"
-#include "menu.h"
+#include "menus/tools_menu.h"
+#include "menus/scene_menu.h"
+#include "menus/settings_menu.h"
 #include "actor.h"
 #include "system.h"
 #include "fifo_queue.h"
