+++ src/menus/hundo_saves_menu.cpp
@@ -1,7 +1,8 @@
 #include "libtp_c/include/tp.h"
 #include "libtp_c/include/system.h"
 #include "libtp_c/include/controller.h"
-#include "menu.h"
+#include "menus/hundo_saves_menu.h"
+#include "menus/practice_menu.h"
 #include "controller.h"
 #include "fifo_queue.h"
 #include "utils/cursor.hpp"
@@ -22,7 +23,6 @@ static CameraMatrix camera = {{0, 0, 0}, {0, 0, 0}};
 static uint16_t angle = 0;
 static Vec3 position = {0, 0, 0};
 bool init_once = false;
-bool hundo_saves_visible;
 
 Line lines[LINES] = {
     {"goats 1", HND_GOATS_1_INDEX, "goat herding 1"},
@@ -114,9 +114,7 @@ void default_load() {
     practice_file.inject_options_after_load = SaveInjector::inject_default_after;
     inject_save_flag = true;
     fifo_visible = true;
-    prac_visible = false;
-    hundo_saves_visible = false;
-    mm_visible = false;
+    MenuRendering::set_menu(MN_NONE_INDEX);
     init_once = false;
 }
 
@@ -311,9 +309,7 @@ void load_save(uint32_t id) {
 
 void HundoSavesMenu::render(Font& font) {
     if (button_is_pressed(Controller::B)) {
-        hundo_saves_visible = false;
-        prac_visible = true;
-        mm_visible = false;
+        MenuRendering::set_menu(MN_PRACTICE_INDEX);
         init_once = false;
         return;
     };
