+++ src/menus/portal_flags_menu.cpp
@@ -1,5 +1,5 @@
 #include "font.h"
-#include "menu.h"
+#include "menus/flags_menu.h"
 #include "controller.h"
 #include "utils/cursor.hpp"
 #include "utils/lines.hpp"
@@ -10,7 +10,6 @@
 
 static Cursor cursor = {0, 0};
 bool init_once = false;
-bool portal_flags_visible;
 
 bool spring_warp;
 bool south_faron_warp;
@@ -66,9 +65,7 @@ void PortalFlagsMenu::render(Font& font) {
 
     if (button_is_pressed(Controller::B)) {
         init_once = false;
-        portal_flags_visible = false;
-        flags_menu_visible = true;
-        mm_visible = false;
+		MenuRendering::set_menu(MN_FLAGS_INDEX);
         return;
     }
